Change how db auth is done for dev and test.
Not specifying a username and password means we can fall back to ident
based auth.  This relies on a postgres role with createdb permissions
existing for the user running.  This has been configured on the dev VM
and on CI.

This change also updates the jenkins script to drop the database before
each run to ensure a clean state.