Change owner of /part-year-profit-tax-credits
This path currently redirects to /end-tax-credits-claim/self-employed. This
redirect was put in place until we were ready to deploy a new Smart Answer at
this path. The new Smart Answer is now ready so we need to "claim" this path.

@alext said that the existing redirect was created by 'short-url-manager' and
that we need to change the owner in url-arbiter to 'smartanswers'.

This migration follows the pattern of a similar migration added in
84ad8957709892d7dcb24459b15fba2f4a590da5.